moving the query file
