+++ clients/core/src/LandingPage/LandingPage.tsx
@@ -4,7 +4,7 @@ import { CourseCard } from './components/CourseCard'
 import { Footer } from './components/Footer'
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
 import { useQuery } from '@tanstack/react-query'
-import { getAllCourses } from '@/network/course'
+import { getAllCourses } from '../network/queries/course'
 import { Course } from '@/interfaces/course'
 
 export function LandingPage(): JSX.Element {
