+++ clients/core/src/management/ManagementConsole.tsx
@@ -8,7 +8,7 @@ import { LoadingPage } from '@/components/LoadingPage'
 import React, { useEffect } from 'react'
 import { useQuery } from '@tanstack/react-query'
 import { Course } from '@/interfaces/course'
-import { getAllCourses } from '@/network/course'
+import { getAllCourses } from '../network/queries/course'
 import { useCourseStore } from '@/zustand/useCourseStore'
 import { ErrorPage } from '@/components/ErrorPage'
 import { Separator } from '@/components/ui/separator'
