+++ clients/core/src/network/queries/course.ts
@@ -3,7 +3,6 @@ import { notAuthenticatedAxiosInstance } from '@/network/configService'
 
 export const getAllCourses = async (): Promise<Course[]> => {
   try {
-    console.log('getCourses')
     return (await notAuthenticatedAxiosInstance.get(`/api/courses/`)).data
   } catch (err) {
     console.error(err)
